Fixed all bugs, maybe

+++ src/ControllerRebalanceThread.java
@@ -102,6 +102,7 @@ class ControllerRebalanceThread implements Runnable
 
                         if(storeVector.size() < R)
                         {
+                            lastRebalance.set(System.currentTimeMillis());
                             storeVectorChangeLock.unlock();
                             continue;
                         }
@@ -150,10 +151,30 @@ class ControllerRebalanceThread implements Runnable
                         storeVector.clear();
                         storeVector.addAll(stableDStores);
 
+                        if(storeVector.size() < R)
+                        {
+                            lastRebalance.set(System.currentTimeMillis());
+                            storeVectorChangeLock.unlock();
+                            continue;
+                        }
+
+                        for(int i=0; i<storeVector.size(); i++)
+                        {
+                            int curDStore = storeVector.get(i).getPort();
+                            Vector<String> actualFiles = RebalanceInfo.currentDStoreList.get(curDStore);
+
+                            if (actualFiles == null) continue;
+                            System.out.println("1 Port " + curDStore + " has " + actualFiles.size() + ", but in the vector it's " + storeVector.get(i).getNumFiles());
+                            //System.out.println(actualFiles.toString());
+                        }
+
+
+
+
                         /** So check if there's any files below or above R, and fill up the numbers by redistributing them randomly,
                          *  and do that by simply starting to build the messages early and including that info there
                          */
-                        ConcurrentHashMap.KeySetView<String, Boolean> inFileIndex_ = fileIndex.newKeySet();
+                        ConcurrentHashMap.KeySetView<String, FileIndex> inFileIndex_ = fileIndex.keySet();
                         String[] inFileIndex = new String[inFileIndex_.size()];
                         inFileIndex_.toArray(inFileIndex);
                         for(int i=0; i<inFileIndex.length; i++)
@@ -171,6 +192,7 @@ class ControllerRebalanceThread implements Runnable
                                         curNum++;
                                         DStoreIndex DStoreTo = storeVector.get(y);
                                         DStoreTo.addFile(file);
+                                        System.out.println("addd1");
                                         int from = RebalanceInfo.currentTotalFileList.get(file).get(new Random().nextInt(RebalanceInfo.currentTotalFileList.get(file).size()));
                                         if(!RebalanceInfo.commandsToSend.containsKey(from))
                                         {
@@ -181,16 +203,19 @@ class ControllerRebalanceThread implements Runnable
                                             tmp2.add(DStoreTo.getPort());
                                             tmp.put(file, tmp2);
                                             RebalanceInfo.commandsToSend.put(from, tmp);
+                                            System.out.println("send10");
                                         }
                                         else if(!RebalanceInfo.commandsToSend.get(from).containsKey(file))
                                         {
                                             Vector<Integer> tmp = new Vector<Integer>();
                                             tmp.add(DStoreTo.getPort());
                                             RebalanceInfo.commandsToSend.get(from).put(file, tmp);
+                                            System.out.println("send9");
                                         }
                                         else
                                         {
                                             RebalanceInfo.commandsToSend.get(from).get(file).add(DStoreTo.getPort());
+                                            System.out.println("send8");
                                         }
 
                                         if(!RebalanceInfo.filesToReceive.containsKey(DStoreTo.getPort()))
@@ -223,10 +248,12 @@ class ControllerRebalanceThread implements Runnable
                                             Vector<String> tmp = new Vector<String>();
                                             tmp.add(file);
                                             RebalanceInfo.commandsToRemove.put(DStoreTo.getPort(), tmp);
+                                            System.out.println("rem8");
                                         }
                                         else
                                         {
                                             RebalanceInfo.commandsToRemove.get(DStoreTo.getPort()).add(file);
+                                            System.out.println("rem7");
                                         }
                                     }
                                 }
@@ -251,7 +278,7 @@ class ControllerRebalanceThread implements Runnable
                             Vector<String> actualFiles = RebalanceInfo.currentDStoreList.get(curDStore);
 
                             if(actualFiles == null) continue;
-                            System.out.println("Port " + curDStore + " has " + actualFiles.size() + ", but in the vector it's " + storeVector.get(i).getNumFiles());
+                            //System.out.println("2 Port " + curDStore + " has " + actualFiles.size() + ", but in the vector it's " + storeVector.get(i).getNumFiles());
                             for(int y=0; y<actualFiles.size(); y++)
                             {
                                 String file = actualFiles.get(y);
@@ -263,10 +290,12 @@ class ControllerRebalanceThread implements Runnable
                                         Vector<String> tmp = new Vector<String>();
                                         tmp.add(file);
                                         RebalanceInfo.commandsToRemove.put(curDStore, tmp);
+                                        System.out.println("rem6");
                                     }
                                     else
                                     {
                                         RebalanceInfo.commandsToRemove.get(curDStore).add(file);
+                                        System.out.println("rem5");
                                     }
                                 }
                             }
@@ -330,8 +359,9 @@ class ControllerRebalanceThread implements Runnable
                          * Think carefully
                          * */
 
-                        storeVectorChangeLock.unlock();
                         lastRebalance.set(System.currentTimeMillis());
+                        storeVectorChangeLock.unlock();
+                        System.out.println("==========================");
                     }
                 }
             }
@@ -342,6 +372,8 @@ class ControllerRebalanceThread implements Runnable
     private void balance()
     {
         storeVector.sort(null);
+
+        //System.out.println("Before balance: " + storeVector.toString());
 
         for(int i=0; i<storeVector.size()/2; i++)
         {
@@ -363,6 +395,8 @@ class ControllerRebalanceThread implements Runnable
 
                     if(!RebalanceInfo.filesToReceive.containsKey(tmpport) && !toSend.contains(file))
                     {
+                        if(lowDStore.getFiles().contains(file)) continue;
+
                         lowDStore.addFile(file);
                         highDStore.removeFile(file);
                         toSend.add(file);
@@ -376,15 +410,66 @@ class ControllerRebalanceThread implements Runnable
                             tmp2.add(lowPort);
                             tmp.put(file,tmp2);
                             RebalanceInfo.commandsToSend.put(highPort, tmp);
+                            System.out.println("send6");
                         }
                         else if(!RebalanceInfo.commandsToSend.get(highPort).containsKey(file))
                         {
                             Vector<Integer> tmp = new Vector<Integer>();
                             tmp.add(lowPort);
                             RebalanceInfo.commandsToSend.get(highPort).put(file, tmp);
+                            System.out.println("send5");
                         }
                         else
                         {
+                            RebalanceInfo.commandsToSend.get(highPort).get(file).add(lowPort);
+                            System.out.println("send4");
+                        }
+
+                        if(!RebalanceInfo.commandsToRemove.containsKey(highPort))
+                        {
+                            Vector<String> tmp = new Vector<String>();
+                            tmp.add(file);
+                            RebalanceInfo.commandsToRemove.put(highPort, tmp);
+                            System.out.println("rem4");
+                        }
+                        else
+                        {
+                            RebalanceInfo.commandsToRemove.get(highPort).add(file);
+                            System.out.println("rem3");
+                        }
+
+                        break;
+                    }
+
+                    else if (!RebalanceInfo.filesToReceive.get(tmpport).contains(file) && !toSend.contains(file))
+                    {
+                        if(lowDStore.getFiles().contains(file)) continue;
+
+                        lowDStore.addFile(file);
+                        highDStore.removeFile(file);
+                        toSend.add(file);
+                        lowNum++;
+                        highNum--;
+
+                        if(!RebalanceInfo.commandsToSend.containsKey(highPort))
+                        {
+                            ConcurrentHashMap<String, Vector<Integer>> tmp = new ConcurrentHashMap<String, Vector<Integer>>();
+                            Vector<Integer> tmp2 = new Vector<Integer>();
+                            tmp2.add(lowPort);
+                            tmp.put(file,tmp2);
+                            RebalanceInfo.commandsToSend.put(highPort, tmp);
+                            System.out.println("send3");
+                        }
+                        else if(!RebalanceInfo.commandsToSend.get(highPort).containsKey(file))
+                        {
+                            Vector<Integer> tmp = new Vector<Integer>();
+                            tmp.add(lowPort);
+                            RebalanceInfo.commandsToSend.get(highPort).put(file, tmp);
+                            System.out.println("send2");
+                        }
+                        else
+                        {
+                            System.out.println("send1");
                             RebalanceInfo.commandsToSend.get(highPort).get(file).add(lowPort);
                         }
 
@@ -393,48 +478,12 @@ class ControllerRebalanceThread implements Runnable
                             Vector<String> tmp = new Vector<String>();
                             tmp.add(file);
                             RebalanceInfo.commandsToRemove.put(highPort, tmp);
+                            System.out.println("rem2");
                         }
                         else
                         {
                             RebalanceInfo.commandsToRemove.get(highPort).add(file);
-                        }
-
-                        break;
-                    }
-
-                    else if (!RebalanceInfo.filesToReceive.get(tmpport).contains(file) && !toSend.contains(file))
-                    {
-                        lowDStore.addFile(file);
-                        highDStore.removeFile(file);
-                        toSend.add(file);
-                        lowNum++;
-                        highNum--;
-
-                        if(!RebalanceInfo.commandsToSend.containsKey(highPort))
-                        {
-                            ConcurrentHashMap<String, Vector<Integer>> tmp = new ConcurrentHashMap<String, Vector<Integer>>();
-                            Vector<Integer> tmp2 = new Vector<Integer>();
-                            tmp2.add(lowPort);
-                            tmp.put(file,tmp2);
-                            RebalanceInfo.commandsToSend.put(highPort, tmp);
-                        }
-                        else if(!RebalanceInfo.commandsToSend.get(highPort).containsKey(file))
-                        {
-                            Vector<Integer> tmp = new Vector<Integer>();
-                            tmp.add(lowPort);
-                            RebalanceInfo.commandsToSend.get(highPort).put(file, tmp);
-                        }
-                        else RebalanceInfo.commandsToSend.get(highPort).get(file).add(lowPort);
-
-                        if(!RebalanceInfo.commandsToRemove.containsKey(highPort))
-                        {
-                            Vector<String> tmp = new Vector<String>();
-                            tmp.add(file);
-                            RebalanceInfo.commandsToRemove.put(highPort, tmp);
-                        }
-                        else
-                        {
-                            RebalanceInfo.commandsToRemove.get(highPort).add(file);
+                            System.out.println("rem1");
                         }
 
                         break;
@@ -442,5 +491,7 @@ class ControllerRebalanceThread implements Runnable
                 }
             }
         }
+
+        //System.out.println("After balance: " + storeVector.toString());
     }
 }
